Rename browser to module
