+++ module/module_test.go
@@ -1,4 +1,4 @@
-package browser
+package module
 
 import (
 	"testing"
