+++ module/registry_test.go
@@ -1,4 +1,4 @@
-package browser
+package module
 
 import (
 	"context"
