+++ register.go
@@ -2,11 +2,11 @@
 package browser
 
 import (
-	"github.com/grafana/xk6-browser/browser"
+	"github.com/grafana/xk6-browser/module"
 
 	k6modules "go.k6.io/k6/js/modules"
 )
 
 func init() {
-	k6modules.Register("k6/x/browser", browser.New())
+	k6modules.Register("k6/x/browser", module.New())
 }
